+++ mmengine/device/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .utlls import get_max_cuda_memory
+from .utils import get_max_cuda_memory
 
 __all__ = ['get_max_cuda_memory']
